BUGFIX: Use local time for column headers in aggregate_by_task_date
When aggregating records by task and date, the key used for the record
is based on the `date()` of the timer's start timestamp.  Since the
timer is retrieved as an aware datetime object in UTC, if the UTC date
is different from the local date, the record was added to the incorrect
time bucket.

This forces a local time conversion on the timestamp before taking the
date, which will correctly categorize the records.